requirements.txt
selenium-wire
python-dotenv
psycopg2-binary
python-telegram-bot
webdriver-manager
